refactor: move platform types
